refactor: サイドバー等をレイアウトに分離
