+++ front-web/src/app/dashboard/page.tsx
@@ -1,23 +1,20 @@
 'use client';
 
-import React, { useState } from 'react';
-import { BarChart, XAxis, YAxis, CartesianGrid, Tooltip, ResponsiveContainer, LineChart, Line, PieChart, Pie, Cell } from 'recharts';
-import { CheckCircle, AlertCircle, TrendingUp, Users, Heart, Brain, Target } from 'lucide-react';
+import React from 'react';
+import { XAxis, YAxis, CartesianGrid, Tooltip, ResponsiveContainer, LineChart, Line, PieChart, Pie, Cell } from 'recharts';
+import { CheckCircle, AlertCircle, Users, Heart, Brain, Target } from 'lucide-react';
 import {
     Box,
     Flex,
     Text,
     Heading,
-    Button,
-    Stack,
     Grid,
     GridItem,
-    Container
+    Container,
+    Stack
 } from '@chakra-ui/react';
 
 const Dashboard = () => {
-    const [activeTab, setActiveTab] = useState('latest-report');
-
     // Sample data for charts
     const healthData = [
         { name: '1月', 健康: 85, やる気: 78, 人間関係: 82 },
@@ -35,290 +32,207 @@ const Dashboard = () => {
         { name: '低い', value: 10, color: '#EF4444' }
     ];
 
-    const MenuItems = [
-        { id: 'latest-report', label: '最新レポート', icon: TrendingUp },
-        { id: 'data-comparison', label: 'データ比較', icon: BarChart }
-    ];
-
-    const bgColor = 'gray.50';
-    const sidebarBg = 'blue.600';
     const cardBg = 'white';
     const textColor = 'gray.800';
     const mutedColor = 'gray.600';
 
     return (
-        <Flex minH="100vh" bg={bgColor}>
-            {/* Sidebar */}
-            <Box
-                w="64"
-                bgGradient={`linear(to-b, ${sidebarBg}, blue.800)`}
-                color="white"
-                shadow="lg"
-                position="relative"
-                backgroundColor="#3182CE"
-            >
-                <Box p={6}>
-                    <Heading size="lg" mb={8}>経営ダッシュボード</Heading>
-                    <Stack gap={2} align="stretch">
-                        {MenuItems.map((item) => {
-                            const Icon = item.icon;
-                            return (
-                                <Button
-                                    key={item.id}
-                                    onClick={() => setActiveTab(item.id)}
-                                    variant={activeTab === item.id ? 'solid' : 'ghost'}
-                                    bg={activeTab === item.id ? 'white' : 'transparent'}
-                                    color={activeTab === item.id ? 'blue.600' : 'blue.100'}
-                                    _hover={{
-                                        bg: activeTab === item.id ? 'white' : 'blue.700',
-                                        color: activeTab === item.id ? 'blue.600' : 'white'
-                                    }}
-                                    justifyContent="flex-start"
-                                    fontWeight="medium"
-                                    borderRadius="lg"
-                                    transition="all 0.2s"
-                                    shadow={activeTab === item.id ? 'md' : 'none'}
-                                >
-                                    <Icon size={20} style={{ marginRight: '8px' }} />
-                                    {item.label}
-                                </Button>
-                            );
-                        })}
-                    </Stack>
-                </Box>
-
-                {/* User Profile */}
-                <Box
-                    position="absolute"
-                    bottom={0}
-                    left={0}
-                    right={0}
-                    p={6}
-                    borderTop="1px"
-                    borderColor="blue.500"
-                >
-                    <Flex alignItems="center" gap={3}>
-                        <Box
-                            w="40px"
-                            h="40px"
-                            bg="blue.400"
-                            borderRadius="full"
-                            display="flex"
-                            alignItems="center"
-                            justifyContent="center"
-                            color="white"
-                            fontWeight="semibold"
-                        >
-                            利
-                        </Box>
-                        <Stack gap={0}>
-                            <Text fontWeight="medium" fontSize="sm">利用者 太郎</Text>
-                            <Text fontSize="xs" color="blue.200">管理者</Text>
-                        </Stack>
-                    </Flex>
-                </Box>
+        <Container maxW="container.xl" p={8} my={8} backgroundColor="white" borderRadius="xl">
+            <Box mb={8}>
+                <Heading size="2xl" color={textColor} mb={2}>最新レポート</Heading>
+                <Text color={mutedColor}>従業員のみなさまの身体・ココロの健康についてのデータです。</Text>
             </Box>
 
-            {/* Main Content */}
-            <Box flex={1} overflowY="auto" backgroundColor="#EBF8FF">
-                <Container maxW="container.xl" p={8} my={8} backgroundColor="white" borderRadius="xl">
-                    <Box mb={8}>
-                        <Heading size="2xl" color={textColor} mb={2}>最新レポート</Heading>
-                        <Text color={mutedColor}>従業員のみなさまの身体・ココロの健康についてのデータです。</Text>
-                    </Box>
+            {/* Health Section */}
+            <Box bg={cardBg} borderRadius="xl" shadow="lg" p={6} mb={8}>
+                <Flex alignItems="center" gap={3} mb={6}>
+                    <Heart color="red.500" size={28} />
+                    <Heading size="xl" color={textColor}>健康</Heading>
+                </Flex>
+                <Box h="80" mb={6}>
+                    <ResponsiveContainer width="100%" height="100%">
+                        <LineChart data={healthData}>
+                            <CartesianGrid strokeDasharray="3 3" stroke="#f0f0f0" />
+                            <XAxis dataKey="name" stroke="#666" />
+                            <YAxis stroke="#666" />
+                            <Tooltip
+                                contentStyle={{
+                                    backgroundColor: '#f8fafc',
+                                    border: '1px solid #e2e8f0',
+                                    borderRadius: '8px'
+                                }}
+                            />
+                            <Line
+                                type="monotone"
+                                dataKey="健康"
+                                stroke="#10B981"
+                                strokeWidth={3}
+                                dot={{ fill: '#10B981', strokeWidth: 2, r: 5 }}
+                            />
+                            <Line
+                                type="monotone"
+                                dataKey="やる気"
+                                stroke="#3B82F6"
+                                strokeWidth={3}
+                                dot={{ fill: '#3B82F6', strokeWidth: 2, r: 5 }}
+                            />
+                            <Line
+                                type="monotone"
+                                dataKey="人間関係"
+                                stroke="#8B5CF6"
+                                strokeWidth={3}
+                                dot={{ fill: '#8B5CF6', strokeWidth: 2, r: 5 }}
+                            />
+                        </LineChart>
+                    </ResponsiveContainer>
+                </Box>
 
-                    {/* Health Section */}
-                    <Box bg={cardBg} borderRadius="xl" shadow="lg" p={6} mb={8}>
-                        <Flex alignItems="center" gap={3} mb={6}>
-                            <Heart color="red.500" size={28} />
-                            <Heading size="xl" color={textColor}>健康</Heading>
+                <Stack gap={4}>
+                    <Text color={mutedColor} lineHeight="tall">
+                        比較年月当時のスコアと比較すると、健康・やる気・人間関係についてのスコアが上昇しています。
+                        特に「中期経営計画」というワードが頻繁に見られることから、経営層のみなさまの想いが従業員のみなさまに届いたのではないかと考えられます。
+                    </Text>
+                    <Text color={mutedColor} lineHeight="tall">
+                        その中でも特に「〇〇」という施策に対するポジティブな意見が多く見られたことあり、この施策を強く推進していくことで、エンゲージメントの向上に繋がることが考えられます。
+                    </Text>
+                </Stack>
+            </Box>
+
+            {/* Good Points */}
+            <Grid templateColumns={{ base: '1fr', md: 'repeat(2, 1fr)' }} gap={6} mb={8}>
+                <GridItem>
+                    <Box
+                        bg="green.50"
+                        borderColor="green.200"
+                        borderWidth={1}
+                        borderRadius="xl"
+                        p={6}
+                        shadow="sm"
+                    >
+                        <Flex alignItems="center" gap={3} mb={4}>
+                            <CheckCircle color="green.600" size={24} />
+                            <Heading size="md" color="green.800">良い点</Heading>
                         </Flex>
-                        <Box h="80" mb={6}>
+                        <Text color="green.700" lineHeight="tall">
+                            比較年月当時のスコアと比較すると、健康・やる気・人間関係についてのスコアが上昇しています。
+                            特に「中期経営計画」というワードが頻繁に見られることから、経営層のみなさまの想いが従業員のみなさまに届いたのではないかと考えられます。
+                            その中でも特に「〇〇」という施策に対するポジティブな意見が多く見られたことあり、この施策を強く推進していくことで、エンゲージメントの向上に繋がることが考えられます。
+                        </Text>
+                    </Box>
+                </GridItem>
+
+                <GridItem>
+                    <Box
+                        bg="red.50"
+                        borderColor="red.200"
+                        borderWidth={1}
+                        borderRadius="xl"
+                        p={6}
+                        shadow="sm"
+                    >
+                        <Flex alignItems="center" gap={3} mb={4}>
+                            <AlertCircle color="red.600" size={24} />
+                            <Heading size="md" color="red.800">改善点</Heading>
+                        </Flex>
+                        <Text color="red.700" lineHeight="tall">
+                            比較年月当時のスコアと比較すると、モチベーションについてのスコアが下降しています。
+                            xxxx.....
+                        </Text>
+                    </Box>
+                </GridItem>
+            </Grid>
+
+            {/* Statistics Cards */}
+            <Grid templateColumns={{ base: '1fr', md: 'repeat(3, 1fr)' }} gap={6} mb={8}>
+                <GridItem>
+                    <Box bg={cardBg} borderRadius="xl" shadow="lg" p={6}>
+                        <Flex justify="space-between" align="start" mb={4}>
+                            <Flex alignItems="center" gap={3}>
+                                <Users color="blue.600" size={24} />
+                                <Text fontWeight="semibold" color={textColor}>総従業員数</Text>
+                            </Flex>
+                            <Text color="blue.600" fontSize="2xl" fontWeight="bold">247</Text>
+                        </Flex>
+                        <Text fontSize="sm" color={mutedColor}>前月比 +3.2%</Text>
+                    </Box>
+                </GridItem>
+
+                <GridItem>
+                    <Box bg={cardBg} borderRadius="xl" shadow="lg" p={6}>
+                        <Flex justify="space-between" align="start" mb={4}>
+                            <Flex alignItems="center" gap={3}>
+                                <Heart color="red.500" size={24} />
+                                <Text fontWeight="semibold" color={textColor}>健康スコア</Text>
+                            </Flex>
+                            <Text color="red.500" fontSize="2xl" fontWeight="bold">94</Text>
+                        </Flex>
+                        <Text fontSize="sm" color={mutedColor}>前月比 +4.2%</Text>
+                    </Box>
+                </GridItem>
+
+                <GridItem>
+                    <Box bg={cardBg} borderRadius="xl" shadow="lg" p={6}>
+                        <Flex justify="space-between" align="start" mb={4}>
+                            <Flex alignItems="center" gap={3}>
+                                <Target color="green.600" size={24} />
+                                <Text fontWeight="semibold" color={textColor}>エンゲージメント</Text>
+                            </Flex>
+                            <Text color="green.600" fontSize="2xl" fontWeight="bold">89</Text>
+                        </Flex>
+                        <Text fontSize="sm" color={mutedColor}>前月比 +2.1%</Text>
+                    </Box>
+                </GridItem>
+            </Grid>
+
+            {/* Motivation Distribution */}
+            <Box bg={cardBg} borderRadius="xl" shadow="lg" p={6}>
+                <Flex alignItems="center" gap={3} mb={6}>
+                    <Brain color="purple.600" size={28} />
+                    <Heading size="xl" color={textColor}>モチベーション分布</Heading>
+                </Flex>
+                <Grid templateColumns={{ base: '1fr', md: 'repeat(2, 1fr)' }} gap={8}>
+                    <GridItem>
+                        <Box h="64">
                             <ResponsiveContainer width="100%" height="100%">
-                                <LineChart data={healthData}>
-                                    <CartesianGrid strokeDasharray="3 3" stroke="#f0f0f0" />
-                                    <XAxis dataKey="name" stroke="#666" />
-                                    <YAxis stroke="#666" />
-                                    <Tooltip
-                                        contentStyle={{
-                                            backgroundColor: '#f8fafc',
-                                            border: '1px solid #e2e8f0',
-                                            borderRadius: '8px'
-                                        }}
-                                    />
-                                    <Line
-                                        type="monotone"
-                                        dataKey="健康"
-                                        stroke="#10B981"
-                                        strokeWidth={3}
-                                        dot={{ fill: '#10B981', strokeWidth: 2, r: 5 }}
-                                    />
-                                    <Line
-                                        type="monotone"
-                                        dataKey="やる気"
-                                        stroke="#3B82F6"
-                                        strokeWidth={3}
-                                        dot={{ fill: '#3B82F6', strokeWidth: 2, r: 5 }}
-                                    />
-                                    <Line
-                                        type="monotone"
-                                        dataKey="人間関係"
-                                        stroke="#8B5CF6"
-                                        strokeWidth={3}
-                                        dot={{ fill: '#8B5CF6', strokeWidth: 2, r: 5 }}
-                                    />
-                                </LineChart>
+                                <PieChart>
+                                    <Pie
+                                        data={motivationData}
+                                        cx="50%"
+                                        cy="50%"
+                                        innerRadius={60}
+                                        outerRadius={100}
+                                        paddingAngle={5}
+                                        dataKey="value"
+                                    >
+                                        {motivationData.map((entry, index) => (
+                                            <Cell key={`cell-${index}`} fill={entry.color} />
+                                        ))}
+                                    </Pie>
+                                    <Tooltip formatter={(value) => [`${value}%`, 'パーセンテージ']} />
+                                </PieChart>
                             </ResponsiveContainer>
                         </Box>
+                    </GridItem>
 
-                        <Stack gap={4}>
-                            <Text color={mutedColor} lineHeight="tall">
-                                比較年月当時のスコアと比較すると、健康・やる気・人間関係についてのスコアが上昇しています。
-                                特に「中期経営計画」というワードが頻繁に見られることから、経営層のみなさまの想いが従業員のみなさまに届いたのではないかと考えられます。
-                            </Text>
-                            <Text color={mutedColor} lineHeight="tall">
-                                その中でも特に「〇〇」という施策に対するポジティブな意見が多く見られたことあり、この施策を強く推進していくことで、エンゲージメントの向上に繋がることが考えられます。
-                            </Text>
+                    <GridItem>
+                        <Stack gap={4} justify="center" h="full">
+                            {motivationData.map((item, index) => (
+                                <Flex key={index} alignItems="center" gap={3} w="full">
+                                    <Box
+                                        w={4}
+                                        h={4}
+                                        borderRadius="full"
+                                        bg={item.color}
+                                    />
+                                    <Text flex={1} color={mutedColor}>{item.name}</Text>
+                                    <Text fontWeight="bold" color={textColor}>{item.value}%</Text>
+                                </Flex>
+                            ))}
                         </Stack>
-                    </Box>
-
-                    {/* Good Points */}
-                    <Grid templateColumns={{ base: '1fr', md: 'repeat(2, 1fr)' }} gap={6} mb={8}>
-                        <GridItem>
-                            <Box
-                                bg="green.50"
-                                borderColor="green.200"
-                                borderWidth={1}
-                                borderRadius="xl"
-                                p={6}
-                                shadow="sm"
-                            >
-                                <Flex alignItems="center" gap={3} mb={4}>
-                                    <CheckCircle color="green.600" size={24} />
-                                    <Heading size="md" color="green.800">良い点</Heading>
-                                </Flex>
-                                <Text color="green.700" lineHeight="tall">
-                                    比較年月当時のスコアと比較すると、健康・やる気・人間関係についてのスコアが上昇しています。
-                                    特に「中期経営計画」というワードが頻繁に見られることから、経営層のみなさまの想いが従業員のみなさまに届いたのではないかと考えられます。
-                                    その中でも特に「〇〇」という施策に対するポジティブな意見が多く見られたことあり、この施策を強く推進していくことで、エンゲージメントの向上に繋がることが考えられます。
-                                </Text>
-                            </Box>
-                        </GridItem>
-
-                        <GridItem>
-                            <Box
-                                bg="red.50"
-                                borderColor="red.200"
-                                borderWidth={1}
-                                borderRadius="xl"
-                                p={6}
-                                shadow="sm"
-                            >
-                                <Flex alignItems="center" gap={3} mb={4}>
-                                    <AlertCircle color="red.600" size={24} />
-                                    <Heading size="md" color="red.800">改善点</Heading>
-                                </Flex>
-                                <Text color="red.700" lineHeight="tall">
-                                    比較年月当時のスコアと比較すると、モチベーションについてのスコアが下降しています。
-                                    xxxx.....
-                                </Text>
-                            </Box>
-                        </GridItem>
-                    </Grid>
-
-                    {/* Statistics Cards */}
-                    <Grid templateColumns={{ base: '1fr', md: 'repeat(3, 1fr)' }} gap={6} mb={8}>
-                        <GridItem>
-                            <Box bg={cardBg} borderRadius="xl" shadow="lg" p={6}>
-                                <Flex justify="space-between" align="start" mb={4}>
-                                    <Flex alignItems="center" gap={3}>
-                                        <Users color="blue.600" size={24} />
-                                        <Text fontWeight="semibold" color={textColor}>総従業員数</Text>
-                                    </Flex>
-                                    <Text color="blue.600" fontSize="2xl" fontWeight="bold">247</Text>
-                                </Flex>
-                                <Text fontSize="sm" color={mutedColor}>前月比 +3.2%</Text>
-                            </Box>
-                        </GridItem>
-
-                        <GridItem>
-                            <Box bg={cardBg} borderRadius="xl" shadow="lg" p={6}>
-                                <Flex justify="space-between" align="start" mb={4}>
-                                    <Flex alignItems="center" gap={3}>
-                                        <Heart color="red.500" size={24} />
-                                        <Text fontWeight="semibold" color={textColor}>健康スコア</Text>
-                                    </Flex>
-                                    <Text color="red.500" fontSize="2xl" fontWeight="bold">94</Text>
-                                </Flex>
-                                <Text fontSize="sm" color={mutedColor}>前月比 +4.2%</Text>
-                            </Box>
-                        </GridItem>
-
-                        <GridItem>
-                            <Box bg={cardBg} borderRadius="xl" shadow="lg" p={6}>
-                                <Flex justify="space-between" align="start" mb={4}>
-                                    <Flex alignItems="center" gap={3}>
-                                        <Target color="green.600" size={24} />
-                                        <Text fontWeight="semibold" color={textColor}>エンゲージメント</Text>
-                                    </Flex>
-                                    <Text color="green.600" fontSize="2xl" fontWeight="bold">89</Text>
-                                </Flex>
-                                <Text fontSize="sm" color={mutedColor}>前月比 +2.1%</Text>
-                            </Box>
-                        </GridItem>
-                    </Grid>
-
-                    {/* Motivation Distribution */}
-                    <Box bg={cardBg} borderRadius="xl" shadow="lg" p={6}>
-                        <Flex alignItems="center" gap={3} mb={6}>
-                            <Brain color="purple.600" size={28} />
-                            <Heading size="xl" color={textColor}>モチベーション分布</Heading>
-                        </Flex>
-                        <Grid templateColumns={{ base: '1fr', md: 'repeat(2, 1fr)' }} gap={8}>
-                            <GridItem>
-                                <Box h="64">
-                                    <ResponsiveContainer width="100%" height="100%">
-                                        <PieChart>
-                                            <Pie
-                                                data={motivationData}
-                                                cx="50%"
-                                                cy="50%"
-                                                innerRadius={60}
-                                                outerRadius={100}
-                                                paddingAngle={5}
-                                                dataKey="value"
-                                            >
-                                                {motivationData.map((entry, index) => (
-                                                    <Cell key={`cell-${index}`} fill={entry.color} />
-                                                ))}
-                                            </Pie>
-                                            <Tooltip formatter={(value) => [`${value}%`, 'パーセンテージ']} />
-                                        </PieChart>
-                                    </ResponsiveContainer>
-                                </Box>
-                            </GridItem>
-
-                            <GridItem>
-                                <Stack gap={4} justify="center" h="full">
-                                    {motivationData.map((item, index) => (
-                                        <Flex key={index} alignItems="center" gap={3} w="full">
-                                            <Box
-                                                w={4}
-                                                h={4}
-                                                borderRadius="full"
-                                                bg={item.color}
-                                            />
-                                            <Text flex={1} color={mutedColor}>{item.name}</Text>
-                                            <Text fontWeight="bold" color={textColor}>{item.value}%</Text>
-                                        </Flex>
-                                    ))}
-                                </Stack>
-                            </GridItem>
-                        </Grid>
-                    </Box>
-                </Container>
+                    </GridItem>
+                </Grid>
             </Box>
-        </Flex>
+        </Container>
     );
 };
